Remove unused backend handlers
