--- reward_kit/mcp_agent/backends/__init__.py
@@ -1,1 +1,0 @@
-# MCP Agent Backends Package
