--- tests/mcp_agent/backends/__init__.py
@@ -1,1 +1,0 @@
-# Tests for MCP Agent Backends
